+++ module/apmgormv2/apmgorm_test.go
@@ -30,9 +30,9 @@ import (
 	"gorm.io/gorm"
 
 	"go.elastic.co/apm/apmtest"
-	mysql "go.elastic.co/apm/module/apmgormv2/dialects/mysql"
-	postgres "go.elastic.co/apm/module/apmgormv2/dialects/postgres"
-	sqlite "go.elastic.co/apm/module/apmgormv2/dialects/sqlite"
+	mysql "go.elastic.co/apm/module/apmgormv2/driver/mysql"
+	postgres "go.elastic.co/apm/module/apmgormv2/driver/postgres"
+	sqlite "go.elastic.co/apm/module/apmgormv2/driver/sqlite"
 	"go.elastic.co/apm/module/apmsql"
 )
